requirements.txt
beautifulsoup4==4.4.1
bs4==0.0.1
cssselect==0.9.1
Django==1.9.6
django-filters==0.2.1
django-rest-swagger==0.3.7
djangorestframework==3.3.3
docopt==0.6.2
pipreqs==0.4.2
PyYAML==3.11
requests==2.10.0
wheel==0.26.0
yarg==0.1.9
